added authentication and bug fixed

+++ backend/routes/taskRoutes.js
@@ -1,13 +1,22 @@
 import express from "express";
 import TaskModel from "../models/TaskModel.js";
-
+import jwt from "jsonwebtoken"
 const router = express.Router();
 
-router.post("/",async(req,res)=>{
+
+
+
+
+
+
+
+
+router.post("/new",async(req,res)=>{
     const {title, description} = req.body;
 
     try {
-        const task = await TaskModel.create({title,description});
+        
+        const task = await TaskModel.create({title,description,userID: req.user.id});
         res.status(200).json(task);
     } catch (error) {
         res.status(400).json({error: error.message});
@@ -16,7 +25,7 @@ router.post("/",async(req,res)=>{
 
 router.get('/',async(req,res)=>{
     try {
-        const task = await TaskModel.find({}).sort({createdAt: -1});
+        const task = await TaskModel.find({userID: req.user.id}).sort({createdAt: -1});
         res.status(200).json(task);
     } catch (error) {
         res.status(400).json({error: error.message});
@@ -27,7 +36,7 @@ router.delete("/:id",async(req,res)=>{
     const {id} = req.params;
 
     try {
-        const task = await TaskModel.findOneAndDelete({_id:id});
+        const task = await TaskModel.findOneAndDelete({_id:id, userID:req.user.id});
         if(!task){
             return res.status(404).json({error: "yeh task nahi hai mila"});
         }
@@ -44,7 +53,9 @@ router.put("/:id",async(req,res)=>{
   try {
       const {status} = req.body;
   
-      const updatedTask = await TaskModel.findByIdAndUpdate(req.params.id, {status:status},{new: true});
+      const updatedTask = await TaskModel.findByIdAndUpdate({
+        _id: req.params.id, userID: req.user.id,
+      }, {status:status},{new: true});
   
       res.json(updatedTask);
     //   console.log(`heyyy ${updatedTask}`);
